LayoutTests/http/tests/security/contentSecurityPolicy/1.1/frame-ancestors/frame-ancestors-nested-same-in-cross-url-block-expected.txt

CONSOLE MESSAGE: Refused to display 'http://localhost:8080/security/contentSecurityPolicy/resources/frame-ancestors.pl?policy=http://127.0.0.1:8000' in a frame because an ancestor violates the following Content Security Policy directive: "frame-ancestors http://127.0.0.1:8000".
A 'frame-ancestors' CSP directive with a URL value should block or allow rendering in nested frames as appropriate.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS The inner IFrame passed.
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Testing a same-origin child with a policy of "http://127.0.0.1:8000" nested in a cross-origin parent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


IFrame load event fired: the IFrame is cross-origin (or was blocked).
PASS The IFrame should have been blocked (or cross-origin). It was.


--------
Frame: '<!--framePath //<!--frame0-->/<!--frame0-->-->'
--------
